backend/requirements.txt

Flask
requests
nltk
Pillow
openai
Celery